Incomplete Report Submitted...
The constraint of limited knowledge required to write a report, well-formatted, with tables and figures, was a limitation I did not foresee in advance. This led to unforeseen delays in compiling the required information for submission. However, albeit a little too late this time to improve, the aim of improving is what will inform the duration of time delegated to future reports written in LaTeX.